pyproject.toml
[project]
name = "mcp-server-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.2",
    "google-generativeai>=0.8.6",
    "groq>=0.31.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "trafilatura>=2.0.0",
]
